+++ src/Telegram.Bot/Requests/Available methods/Messages/Location/EditMessageLiveLocationRequest.cs
@@ -53,7 +53,7 @@ public class EditMessageLiveLocationRequest : RequestBase<Message>, IChatTargeta
     /// </summary>
     [JsonInclude]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public float? HorizontalAccuracy { get; set; }
+    public double? HorizontalAccuracy { get; set; }
 
     /// <summary>
     /// Direction in which the user is moving, in degrees. Must be between 1 and 360 if specified.
